+++ clientui/src/main/java/com/projet9/clientui/Dto/AddressDto.java
@@ -1,6 +1,5 @@
-package com.projet9.clientui.controller.Dto;
+package com.projet9.clientui.Dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,16 +9,14 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class GenderDto {
+public class AddressDto {
 
-    int id;
-
-    private String gender;
+    private String number;
+    private String street;
 
     @Override
     public String toString()
     {
-        return gender;
+        return number + " " + street;
     }
 }
-
