+++ clientui/src/main/java/com/projet9/clientui/controller/ClientController.java
@@ -1,9 +1,11 @@
 package com.projet9.clientui.controller;
 
+import com.projet9.clientui.Dto.PatientDto;
 import com.projet9.clientui.proxies.PatientServiceProxy;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 public class ClientController {
@@ -17,6 +19,17 @@ public class ClientController {
     @GetMapping("/patient")
     public String home (Model model) {
         model.addAttribute("patients", patientProxy.getPatients());
-        return "patient";
+        return "patient-list";
+    }
+
+    @GetMapping("/patient/add")
+    public String showAddNewPatientForm (Model model) {
+        model.addAttribute("patient", new PatientDto());
+        return "add-new-patient";
+    }
+
+    @PostMapping("/patient/validate")
+    public void validatePatient (Model model, @ModelAttribute @DateTimeFormat(pattern = "dd/MM/yyyy") PatientDto patientDto) {
+        System.out.println(patientDto.getFirstName()+ " ici " + patientDto.getDateOfBirth());
     }
 }
