+++ clientui/src/main/resources/templates/patient-list.html
