+++ patient-service/src/main/java/com/projet9/patientservice/model/Patient.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.Date;
+import java.time.LocalDate;
 
 @Getter
 @Setter
@@ -27,7 +27,8 @@ public class Patient {
     @Column(name = "last_name")
     private String lastName;
     @Column(name = "date_of_birth")
-    private Date dateOfBirth;
+
+    private LocalDate dateOfBirth;
 
     @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "gender_id")
